Move StatsD singleton methods into a legacy_client, to prepare for a swappable client.
